messing around with this still. It compiles again, so I'm saving.
It crashes when running the compiled jar, but from the repl this works again. I need to update this so we're not using the old variables, and I have the function working to read in all of the files in a directory but it's not working when firing on the UI Event and I'm still not sure why. Also the fn-fx library still doesn't have an implementation for the file chooser in JavaFX, so I may have to build a lil wonky work-around for this after I get the directory selection code working. Having fun so far, lol.